角色管理和菜单管理

+++ fnsco-service-auth/src/main/java/net/fnsco/auth/service/sys/entity/RoleDeptDO.java
@@ -6,41 +6,41 @@ public class RoleDeptDO {
     /**
      * 
      */
-    private Integer id;
+    private Long id;
 
     /**
      * 角色ID
      */
-    private Integer roleId;
+    private Long roleId;
 
     /**
      * 部门ID
      */
-    private Integer deptId;
+    private Long deptId;
 
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
 
-    public Integer getDeptId() {
+    public Long getDeptId() {
         return deptId;
     }
 
-    public void setDeptId(Integer deptId) {
+    public void setDeptId(Long deptId) {
         this.deptId = deptId;
     }
